+++ navigation/side-navigation/src/common/constants.ts
@@ -8,3 +8,5 @@ export const itemTextColor = N500;
 export const itemTextSelectedColor = B400;
 
 export const separatorColor = N30;
+export const VAR_SEPARATOR_COLOR = '--ds-menu-seperator-color';
+export const VAR_SCROLL_INDICATOR_COLOR = '--ds-menu-scroll-indicator-color';