+++ confluence/legacy-mobile-macros/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @atlaskit/legacy-mobile-macros
 
+## 0.3.3
+
+### Patch Changes
+
+- Updated dependencies
+
 ## 0.3.2
 
 ### Patch Changes